# docs/swagger.yaml
definitions:
...
  university_system_internal_domain_models.Course:
    properties:
      code:
        type: string
      created_at:
        type: string
      credits:
        type: integer
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      teacher_id:
        type: integer
      updated_at:
        type: string
    type: object
  university_system_internal_domain_models.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  university_system_internal_domain_models.Teacher:
    properties:
      birthdate:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      department:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      password:
        type: string
      position:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  university_system_internal_domain_models.User:
    properties:
      birthdate:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: n_rakhat@kbtu.kz
    name: Rakhat Nurtileu
    url: http://linkedin.com/in/nurtileu-rakhat-02aba3294/
  description: System для управления пользователями, курсами и ролями
  title: University System
  version: "1.0"
paths:
  /api/users:
    get:
      consumes:
      - application/json
      description: Возвращает всех пользователей системы
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/university_system_internal_domain_models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      summary: Создание пользователя
      tags:
      - users
  /api/users/{id}:
    delete:
      description: Удаляет пользователя по его ID
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Пользователь успешно удалён
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление пользователя
      tags:
      - users
    get:
      consumes:
      - application/json
      description: get user by id
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: get user by id
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет информацию о пользователе по его ID
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.User'
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление данных пользователя
      tags:
      - users
...
  /teachers:
    get:
      consumes:
      - application/json
      description: Возвращает список всех преподавателей в системе
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/university_system_internal_domain_models.Teacher'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список всех преподавателей
      tags:
      - teachers
    post:
      consumes:
      - application/json
      description: Добавляет нового преподавателя в систему
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные нового преподавателя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.Teacher'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Teacher'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать нового преподавателя
      tags:
      - teachers
  /teachers/{id}:
    delete:
      description: Удаляет преподавателя по его ID
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID преподавателя
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить преподавателя
      tags:
      - teachers
    get:
      consumes:
      - application/json
      description: Возвращает данные преподавателя по его ID
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID преподавателя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Teacher'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить информацию о преподавателе
      tags:
      - teachers
    put:
      consumes:
      - application/json
      description: Обновляет информацию о преподавателе
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID преподавателя
        in: path
        name: id
        required: true
        type: string
      - description: Обновленные данные преподавателя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.Teacher'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Teacher'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить данные преподавателя
      tags:
      - teachers
  /teachers/{id}/courses:
    get:
      consumes:
      - application/json
      description: Возвращает список курсов, которые ведет преподаватель
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID преподавателя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/university_system_internal_domain_models.Course'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить курсы преподавателя
      tags:
      - teachers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"